requirements.txt
# AI & Agent Framework
pydantic-ai>=0.0.14
pydantic>=2.0.0,<3.0.0

# Windows Desktop Automation
pynput>=1.7.6
pywin32>=306
pywinauto>=0.6.8

# Template Engine & Configuration
jinja2>=3.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Browser & Web (Optional - for DOM inspection)
playwright>=1.40.0

# HTTP Client (for potential API calls)
httpx>=0.27.0

# Logging & Utilities
colorama>=0.4.6
